+++ src/tasks.h
@@ -373,8 +373,6 @@ enum SEND_TRANSACTION_ETH {
 enum SIGN_MESSAGE_ETH {
   SIGN_MSG_VERIFY_COIN_ETH = 1,
   SIGN_MSG_RAW_MSG_WAIT_SCREEN_ETH,
-  SIGN_MSG_RAW_MSG_RECEIVED_ETH,
-  SIGN_MSG_VERIFY_CONTRACT_ADDRESS_ETH,
   SIGN_MSG_DISPLAY_INFO_ETH,
   SIGN_MSG_CHECK_PASSPHRASE_ETH,
   SIGN_MSG_ENTER_PASSPHRASE_ETH,
